+++ lib/src/main/java/com/freshdigitable/yttt/AddFreeTalkWorker.kt
@@ -9,6 +9,7 @@ import androidx.work.Data
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
+import com.freshdigitable.yttt.data.YouTubeFacade
 import com.freshdigitable.yttt.data.model.YouTubeVideo
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
