+++ lib/src/main/java/com/freshdigitable/yttt/FetchStreamUseCase.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import com.freshdigitable.yttt.data.AccountRepository
 import com.freshdigitable.yttt.data.SettingRepository
 import com.freshdigitable.yttt.data.TwitchLiveRepository
+import com.freshdigitable.yttt.data.YouTubeFacade
 import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.YouTubeSubscriptionSummary
 import com.freshdigitable.yttt.data.model.YouTubeSubscriptionSummary.Companion.needsUpdatePlaylist
@@ -87,54 +88,6 @@ class FetchYouTubeStreamUseCase @Inject constructor(
     companion object {
         @Suppress("unused")
         private val TAG = FetchYouTubeStreamUseCase::class.simpleName
-    }
-}
-
-class YouTubeFacade @Inject constructor(
-    private val repository: YouTubeRepository,
-) {
-    suspend fun fetchVideoList(ids: Collection<YouTubeVideo.Id>): List<YouTubeVideo> {
-        val videos = repository.fetchVideoList(ids)
-        val unchecked = videos.filter { it.isFreeChat == null }
-            .associateWith { isFreeChat(it) }
-        val add = unchecked.entries.filter { (_, f) -> f }.map { it.key.id }
-        val remove = unchecked.map { it.key.id } - add.toSet()
-        if (add.isNotEmpty()) {
-            repository.addFreeChatItems(add)
-        }
-        if (remove.isNotEmpty()) {
-            repository.removeFreeChatItems(remove)
-        }
-        return repository.fetchVideoList(ids)
-    }
-
-    private fun isFreeChat(video: YouTubeVideo): Boolean {
-        return regex.any { video.title.contains(it) }
-    }
-
-    suspend fun updateAsFreeChat() {
-        val unchecked = repository.findAllUnfinishedVideos()
-        val freeChat = unchecked.filter { it.isFreeChat == null }
-            .filter { v -> regex.any { v.title.contains(it) } }
-            .map { it.id }
-        repository.addFreeChatItems(freeChat)
-        val unfinished = unchecked.filter { it.isFreeChat == null }.map { it.id } - freeChat.toSet()
-        repository.removeFreeChatItems(unfinished)
-    }
-
-    suspend fun addFreeChatFromWorker(id: YouTubeVideo.Id) {
-        val v = repository.fetchVideoList(listOf(id))
-        repository.addFreeChatItems(v.map { it.id })
-    }
-
-    companion object {
-        private val regex = listOf(
-            "free chat".toRegex(RegexOption.IGNORE_CASE),
-            "フリーチャット".toRegex(),
-            "ふりーちゃっと".toRegex(),
-            "schedule".toRegex(RegexOption.IGNORE_CASE),
-            "の予定".toRegex(),
-        )
     }
 }
